Converge on one budget model and fix integration breakers
Three budget implementations had accumulated: MonthlyBudget/CategoryBudget
(new, FK-based), the older Budget (category as a name string), and the
Expense entity the old budget summaries read from. The frontend spoke only
to the old one, so the new controllers were unreachable and the summaries
could only ever report zero.

Budgets now live entirely in MonthlyBudget + CategoryBudget:

- Spend comes from the transaction ledger via a new
  util/BudgetSpendCalculator, which wraps the same
  getCategoryWiseBreakdown query the dashboard uses, so the two cannot
  disagree about a month. The old path read a separate `expenses` table
  that no controller ever wrote to.
- Caps join to spend on categoryId rather than a category name, so
  renaming a category can no longer orphan its budget.
- Adds the missing pieces the new model needed to be usable: `id` on both
  responses (nothing could be deleted without it), DELETE for each
  resource with an ownership check, and summary endpoints at
  /budgets/monthly/summary and /budgets/category/summary.
- GET /budgets/monthly returns 200 with an empty body when no cap is set.
  It previously threw a not-found that surfaced as 400, which is the
  normal state on a fresh account, not an error.
- Deletes the old Budget and Expense stacks. Their tables are dropped; the
  one real budget row was migrated into monthly_budgets first.

Alongside that, the integration fixes that were blocked on this:

- SecurityConfig had no exceptionHandling, so Spring fell back to
  Http403ForbiddenEntryPoint and an expired token answered 403 with an
  empty body — while the axios interceptor keys off 401 to clear the token
  and redirect. Every screen silently broke once the token aged out. Adds
  a 401 entry point and 403 denied handler emitting the same
  {"message": ...} body the controllers use, guards the interceptor
  against redirect loops on /login, and has ProtectedRoute check the exp
  claim so an expired session bounces before rendering.
- JWT secret and expiry were hardcoded in JwtConfig while
  application.yml advertised a 24-hour expiry nothing read; tokens
  actually died after an hour. JwtTokenProvider now injects both
  properties and JwtConfig is deleted.
- registerUser had no duplicate-email guard (the check sat in an unused
  register() overload), so a repeat signup showed a raw PostgreSQL
  constraint message in the form. AuthServiceImpl.getCurrentUserId now
  delegates to AuthUtil instead of re-reading SecurityContextHolder, so
  there is one implementation of "who is logged in".
- GlobalExceptionHandler built bodies with Map.of, which rejects null
  values, so any exception with a null message made the handler itself
  throw. Missing required params now return 400 rather than 500, and
  unreadable bodies return a fixed message instead of echoing Jackson
  internals.

Schema migration applied to dev (INSERT ... SELECT into monthly_budgets,
then DROP TABLE budgets, budget, expenses). No data lost.

Verified against docker compose: register dedup, 401 on missing/garbage
token, 400 on missing params, monthly and category upsert idempotence,
the sum(category caps) <= monthly cap rule at its boundary, non-zero
budget spend agreeing with the dashboard breakdown, and cross-user
read/write/delete isolation on every budget route. The old
/budgets/user/{userId}/... paths now 404. Frontend lints and builds.

+++ backend/src/main/java/com/omkar/expensetracker/dto/response/BudgetSummaryResponse.java
@@ -2,6 +2,8 @@ package com.omkar.expensetracker.dto.response;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -9,9 +11,9 @@ import lombok.*;
 @Builder
 public class BudgetSummaryResponse {
 
-    private Double budget;          // Total budget
-    private Double spent;           // Total expenses
-    private Double remaining;       // budget - spent
-    private Double percentageUsed;  // (spent/budget)*100
-    private String status;          // OVER_BUDGET or UNDER_BUDGET
+    private BigDecimal budget;          // Monthly cap, 0 when none is set
+    private BigDecimal spent;           // Total EXPENSE transactions in the month
+    private BigDecimal remaining;       // budget - spent
+    private BigDecimal percentageUsed;  // (spent/budget)*100
+    private String status;              // OVER_BUDGET or UNDER_BUDGET
 }
